refactor: rename Space packages to the @endo/space-* convention
Align the Space packages with the naming used by space-file-explorer:

- @endo/whylip            -> @endo/space-whylip
- @endo/chat-network-view -> @endo/space-peers
- @endo/inventory-graph   -> @endo/space-inventory-graph

Directories moved with git mv (history preserved); package names,
workspace dependencies in @endo/chat, CSS side-effect imports and
component import specifiers, the tsconfig/typedoc exclude paths, and the
referencing design docs are all updated. The yarn.lock relink lands in a
separate commit.
