+++ packages/chat/package.json
@@ -48,7 +48,6 @@
   },
   "dependencies": {
     "@endo/captp": "workspace:^",
-    "@endo/chat-network-view": "workspace:^",
     "@endo/daemon": "workspace:^",
     "@endo/eventual-send": "workspace:^",
     "@endo/exo": "workspace:^",
@@ -56,7 +55,6 @@
     "@endo/far": "workspace:^",
     "@endo/harden": "workspace:^",
     "@endo/init": "workspace:^",
-    "@endo/inventory-graph": "workspace:^",
     "@endo/markmdown": "workspace:^",
     "@endo/monaco-wrapper": "workspace:^",
     "@endo/pass-style": "workspace:^",
@@ -65,8 +63,10 @@
     "@endo/preact-container": "workspace:^",
     "@endo/promise-kit": "workspace:^",
     "@endo/space-file-explorer": "workspace:^",
+    "@endo/space-inventory-graph": "workspace:^",
+    "@endo/space-peers": "workspace:^",
+    "@endo/space-whylip": "workspace:^",
     "@endo/where": "workspace:^",
-    "@endo/whylip": "workspace:^",
     "marked": "^17.0.5",
     "monaco-editor": "^0.55.1",
     "preact": "^10.29.2",
