+++ packages/space-inventory-graph/index.js
