+++ packages/space-inventory-graph/src/graph.css
