+++ packages/space-inventory-graph/src/graph.js
