+++ packages/space-inventory-graph/src/mount.js
