+++ packages/space-inventory-graph/tsconfig.json
