+++ packages/space-peers/tsconfig.json
