+++ packages/space-whylip/src/SceneCanvas.js
