+++ tsconfig.json
@@ -16,7 +16,7 @@
         "packages/test262-runner",
         "packages/ses/types.test-d.ts",
         "packages/chat",
-        "packages/chat-network-view",
+        "packages/space-peers",
         "packages/conversation-tree",
         "packages/demo",
         "packages/fae",
@@ -25,7 +25,7 @@
         "packages/eventual-send-test",
         "packages/harden-test",
         "packages/hex-test",
-        "packages/inventory-graph",
+        "packages/space-inventory-graph",
         "packages/ses-test",
         "packages/jaine",
         "packages/lal",
@@ -33,7 +33,7 @@
         "packages/platform",
         "packages/preact-container",
         "packages/space-file-explorer",
-        "packages/whylip",
+        "packages/space-whylip",
         "**/build",
         "**/demo",
         "**/dist",
