+++ typedoc.json
@@ -18,7 +18,7 @@
     "**/node_modules/**",
     "packages/chacha12-fast-check-test",
     "packages/chat",
-    "packages/chat-network-view",
+    "packages/space-peers",
     "packages/cjs-module-analyzer",
     "packages/cli",
     "packages/compartment-mapper",
@@ -32,7 +32,7 @@
     "packages/genie",
     "packages/harden-test",
     "packages/hex-test",
-    "packages/inventory-graph",
+    "packages/space-inventory-graph",
     "packages/ses-test",
     "packages/jaine",
     "packages/hardened262",
@@ -50,7 +50,7 @@
     "packages/syrup",
     "packages/test262-runner/",
     "packages/where",
-    "packages/whylip",
+    "packages/space-whylip",
     "packages/monaco-wrapper",
     "packages/space-file-explorer",
     "api-docs/"
